rename functional tests
