+++ tests/functional/test_memory.py
